# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/download/{filename}": {
            "get": {
                "description": "Faz o download de um arquivo pelo nome",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Video"
                ],
                "summary": "Download de arquivo processado",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome do arquivo",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Arquivo para download",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Arquivo não encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/health": {
            "get": {
                "description": "Retorna o status do serviço",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Status da API",
                "responses": {
                    "200": {
                        "description": "Health check",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/upload": {
            "post": {
                "description": "Recebe um arquivo de vídeo para processamento",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Video"
                ],
                "summary": "Upload de vídeo",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Arquivo de vídeo",
                        "name": "video",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Upload realizado com sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Erro no upload",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    }
}